+++ server/src/main/java/com/zmission/codespot/util/CourseUtil.java
@@ -1,3 +1,5 @@
+package com.zmission.codespot.util;
+
 public class CourseUtil {
 
     private CourseUtil() {}
